+++ packages/pangea-sdk/src/test/java/cloud/pangeacyber/pangea/ConfigTest.java
@@ -19,7 +19,7 @@ public class ConfigTest {
 		assertNotNull(config.getDomain());
 		assertNotNull(config.getToken());
 		assertNotNull(config.getConnectionTimeout());
-		assertNotNull(config.getEnviroment());
+		assertNotNull(config.getEnvironment());
 		assertNotNull(config.isInsecure());
 		assertEquals(50, config.getMaxTotalConnections());
 		assertEquals(50, config.getMaxConnectionsPerRoute());
@@ -27,7 +27,7 @@ public class ConfigTest {
 
 	@Test
 	public void testInsecureTrueEnvironmentLocal() throws ConfigException {
-		Config config = new Config.Builder(token, domain).insecure(true).enviroment("local").build();
+		Config config = new Config.Builder(token, domain).insecure(true).environment("local").build();
 		String url = config.getServiceUrl(serviceName, path).toString();
 		String urlExpected = new StringBuilder().append("http://").append(domain).append(path).toString();
 		assertEquals(urlExpected, url);
@@ -35,7 +35,7 @@ public class ConfigTest {
 
 	@Test
 	public void testInsecureFalseEnvironmentLocal() throws ConfigException {
-		Config config = new Config.Builder(token, domain).insecure(false).enviroment("local").build();
+		Config config = new Config.Builder(token, domain).insecure(false).environment("local").build();
 		String url = config.getServiceUrl(serviceName, path).toString();
 		String urlExpected = new StringBuilder().append("https://").append(domain).append(path).toString();
 		assertEquals(urlExpected, url);
@@ -43,7 +43,7 @@ public class ConfigTest {
 
 	@Test
 	public void testInsecureTrueEnvironmentProduction() throws ConfigException {
-		Config config = new Config.Builder(token, domain).insecure(true).enviroment("production").build();
+		Config config = new Config.Builder(token, domain).insecure(true).environment("production").build();
 		String url = config.getServiceUrl(serviceName, path).toString();
 		String urlExpected = new StringBuilder()
 			.append("http://")
@@ -57,7 +57,7 @@ public class ConfigTest {
 
 	@Test
 	public void testInsecureFalseEnvironmentProduction() throws ConfigException {
-		Config config = new Config.Builder(token, domain).insecure(false).enviroment("production").build();
+		Config config = new Config.Builder(token, domain).insecure(false).environment("production").build();
 		String url = config.getServiceUrl(serviceName, path).toString();
 		String urlExpected = new StringBuilder()
 			.append("https://")