+++ packages/pangea-sdk/src/test/java/cloud/pangeacyber/pangea/file_scan/ITFileScanTest.java
@@ -9,6 +9,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import cloud.pangeacyber.pangea.Config;
 import cloud.pangeacyber.pangea.FileData;
 import cloud.pangeacyber.pangea.Helper;
+import cloud.pangeacyber.pangea.SkipAccepted;
 import cloud.pangeacyber.pangea.TestEnvironment;
 import cloud.pangeacyber.pangea.TransferMethod;
 import cloud.pangeacyber.pangea.Utils;
@@ -65,6 +66,7 @@ public class ITFileScanTest {
 	}
 
 	@Test
+	@SkipAccepted
 	public void testFileScan_Scan_multipart() throws PangeaException, PangeaAPIException, IOException {
 		File file = new File(TESTFILE_PATH);
 		FileScanResponse response = client.scan(
@@ -137,6 +139,7 @@ public class ITFileScanTest {
 	}
 
 	@Test
+	@SkipAccepted
 	public void testFileScan_Scan_reversinglabs() throws PangeaException, PangeaAPIException, IOException {
 		File file = new File(TESTFILE_PATH);
 		FileScanResponse response = client.scan(
@@ -202,8 +205,8 @@ public class ITFileScanTest {
 				assertNull(response.getResult().getParameters());
 				assertNotNull(response.getResult().getRawData());
 				break;
-			} catch (PangeaAPIException e) {
-				assertTrue(retry < maxRetry - 1);
+			} catch (AcceptedRequestException e) {
+				// No-op.
 			}
 		}
 	}
@@ -247,8 +250,8 @@ public class ITFileScanTest {
 				assertNull(response.getResult().getParameters());
 				assertNotNull(response.getResult().getRawData());
 				break;
-			} catch (PangeaAPIException e) {
-				assertTrue(retry < maxRetry - 1);
+			} catch (AcceptedRequestException e) {
+				// No-op.
 			}
 		}
 	}
@@ -290,8 +293,8 @@ public class ITFileScanTest {
 				assertNull(response.getResult().getParameters());
 				assertNotNull(response.getResult().getRawData());
 				break;
-			} catch (PangeaAPIException e) {
-				assertTrue(retry < maxRetry - 1);
+			} catch (AcceptedRequestException e) {
+				// No-op.
 			}
 		}
 	}